Status-statement parentheses no longer required (#34)
* Refactor comments and structure of if-feature

* Remove means-statement parens from grammar

The means-statement may have a parentheses, but it is only a part of the expression, not a requirement like in an if-statement.

* Update features to test presence and absence of means-statement parentheses

* Removed parens requirement from status statements

Not all tests are working. In doing this I uncovered a bug in the generation of KnowledgeBase code that appends an extra untranslated equal test, like: = 'a

The grammars were also changed from ascii to UTF-8.

* Complete removal of status parentheses

Left in println's for debugging other things.